Dodanie autentykacji do wszystkich endpointów (#84)

+++ KasaWGrupieBackend/Controllers/UsersController.cs
@@ -13,6 +13,7 @@ namespace KasaWGrupie.API.Controllers;
 
 [Route("users")]
 [ApiController]
+[FirebaseAuthorize]
 public class UsersController(
 	IAuthService authService,
 	IMediator mediator
@@ -24,7 +25,6 @@ public class UsersController(
 	/// <returns>Successfully inserted new User</returns>
 	[HttpPost]
 	[TranslateResultToActionResult]
-	[FirebaseAuthorize]
 	[Consumes("multipart/form-data")]
 	public async Task<Result> CreateUser(IFormCollection formCollection, IFormFile? profilePicture)
 	{
@@ -70,6 +70,10 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result> DeleteUser(int id)
 	{
+		if (id != await authService.GetUserIdFromAuthTokenAsync(HttpContext))
+		{
+			return Result.Forbidden("Cannot delete another user's account.");
+		}
 		var command = new DeleteUserCommand(id);
 		return await mediator.Send(command);
 	}
@@ -78,6 +82,10 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result> UpdateUserName(int id, [FromBody] UpdateUserNameDto updateUserNameDto)
 	{
+		if (id != await authService.GetUserIdFromAuthTokenAsync(HttpContext))
+		{
+			return Result.Forbidden("Cannot update another user's name.");
+		}
 		var command = new UpdateUserNameCommand(id, updateUserNameDto);
 		return await mediator.Send(command);
 	}
@@ -86,6 +94,10 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result> UpdateUserProfilePicture(int id, [FromForm] UpdateUserProfilePictureDto updateUserProfilePictureDto)
 	{
+		if (id != await authService.GetUserIdFromAuthTokenAsync(HttpContext))
+		{
+			return Result.Forbidden("Cannot update another user's profile picture.");
+		}
 		var command = new UpdateUserProfilePictureCommand(id, updateUserProfilePictureDto);
 		return await mediator.Send(command);
 	}
@@ -102,6 +114,7 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result<ICollection<GetUserDto>>> GetFriends(int userId)
 	{
+		// inni użytkownicy mogą czy nie?
 		var command = new GetFriendsCommand(userId);
 		return await mediator.Send(command);
 	}
@@ -110,7 +123,8 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result> AddFriendRequest([FromBody] AddFriendRequestDto addFriendRequestDto)
 	{
-		var command = new AddFriendRequestCommand(addFriendRequestDto);
+		var userId = await authService.GetUserIdFromAuthTokenAsync(HttpContext);
+		var command = new AddFriendRequestCommand(userId, addFriendRequestDto);
 		var result = await mediator.Send(command);
 		return result;
 	}
@@ -119,7 +133,8 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result> ChangeFriendRequestStatus(int requestId, [FromBody] ChangeFriendRequestStatusDto changeFriendRequestStatusDto)
 	{
-		var command = new ChangeFriendRequestStatusCommand(requestId, changeFriendRequestStatusDto);
+		var userId = await authService.GetUserIdFromAuthTokenAsync(HttpContext);
+		var command = new ChangeFriendRequestStatusCommand(userId, requestId, changeFriendRequestStatusDto);
 		var result = await mediator.Send(command);
 		return result;
 	}
@@ -128,6 +143,10 @@ public class UsersController(
 	[TranslateResultToActionResult]
 	public async Task<Result<ICollection<FriendRequestDisplayDto>>> GetFriendRequests(int userId)
 	{
+		if (userId != await authService.GetUserIdFromAuthTokenAsync(HttpContext))
+		{
+			return Result.Forbidden("Cannot get friend requests of another user.");
+		}
 		var command = new GetFriendRequestsCommand(userId);
 		var result = await mediator.Send(command);
 		return result;
